Restructured code (changed packages and package names, and what classes are in the different packages)
