+++ .gitignore
@@ -25,4 +25,4 @@ target
 
 # virtual machine crash logs, see http://www.java.com/en/download/help/error_hotspot.xml
 hs_err_pid*
-/src/main/tex/oldtex/
+/src/main/resources/tex/oldtex/
