+++ src/main/java/inf112/skeleton/app/cards/ICards.java
@@ -1,4 +1,4 @@
-package inf112.skeleton.app.Cards;
+package inf112.skeleton.app.cards;
 
 /**
  * Interface for all cards.
