+++ src/main/java/inf112/skeleton/app/sprites/Direction.java
@@ -1,4 +1,4 @@
-package inf112.skeleton.app.Sprites;
+package inf112.skeleton.app.sprites;
 
 public enum Direction{
     NORTH,
