+++ src/main/java/inf112/skeleton/app/sprites/Wall.java
@@ -1,4 +1,4 @@
-package inf112.skeleton.app.Sprites;
+package inf112.skeleton.app.sprites;
 
 public class Wall extends AbstractGameObject {
     public Wall(String texturePath) {
