+++ src/main/resources/tex/flag1.png
