+++ src/main/resources/tex/flag2.png
