+++ src/main/resources/tex/flag3.png
