+++ src/main/resources/tex/ground.png
