+++ src/main/resources/tex/hole.png
