+++ src/main/resources/tex/wall.png
