+++ src/main/resources/tex/win.png
